+++ src/commands/createWebApp/createWebApp.ts
@@ -5,10 +5,10 @@
 
 import { ConfigurationTarget, MessageItem, workspace, WorkspaceConfiguration } from 'vscode';
 import { AzureParentTreeItem, IActionContext, parseError } from "vscode-azureextensionui";
-import { configurationSettings, extensionPrefix } from "../constants";
-import { SubscriptionTreeItem } from '../explorer/SubscriptionTreeItem';
-import { WebAppTreeItem } from "../explorer/WebAppTreeItem";
-import { ext } from "../extensionVariables";
+import { configurationSettings, extensionPrefix } from "../../constants";
+import { SubscriptionTreeItem } from '../../explorer/SubscriptionTreeItem';
+import { WebAppTreeItem } from "../../explorer/WebAppTreeItem";
+import { ext } from "../../extensionVariables";
 
 export async function createWebApp(context: IActionContext, node?: AzureParentTreeItem | undefined): Promise<void> {
     if (!node) {
@@ -35,5 +35,5 @@ export async function createWebApp(context: IActionContext, node?: AzureParentTr
         throw error;
     }
 
-    newSite.showCreatedOutput(context);
+    newSite.promptToDeploy(context);
 }